#48 moved integration tests in separate package
