+++ src/test/java/juja/microservices/BaseIntegrationTest.java
@@ -1,4 +1,4 @@
-package juja.microservices.gamification;
+package juja.microservices;
 
 import com.github.fakemongo.Fongo;
 import com.lordofthejars.nosqlunit.mongodb.MongoDbConfiguration;
@@ -6,13 +6,14 @@ import com.lordofthejars.nosqlunit.mongodb.MongoDbRule;
 import com.lordofthejars.nosqlunit.mongodb.SpringMongoDbRule;
 import com.mongodb.MockMongoClient;
 import com.mongodb.MongoClient;
+import juja.microservices.gamification.Gamification;
 import org.junit.Rule;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.web.context.WebApplicationContext;
 
 import javax.inject.Inject;
 
-import static juja.microservices.gamification.GamificationTestConfig.TEST_DATABASE_NAME;
+import static juja.microservices.GamificationTestConfig.TEST_DATABASE_NAME;
 
 /**
  * @author danil.kuznetsov
