+++ src/test/java/juja/microservices/gamification/dao/AchievementRepositorySpeedTest.java
@@ -3,7 +3,7 @@ package juja.microservices.gamification.dao;
 import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
 import juja.microservices.gamification.entity.Achievement;
 import juja.microservices.gamification.entity.AchievementType;
-import juja.microservices.gamification.BaseIntegrationTest;
+import juja.microservices.BaseIntegrationTest;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
