+++ src/test/java/juja/microservices/integration/AchievementControllerIntegrationTest.java
@@ -1,7 +1,7 @@
-package juja.microservices.gamification.controller;
+package juja.microservices.integration;
 
 import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
-import juja.microservices.gamification.BaseIntegrationTest;
+import juja.microservices.BaseIntegrationTest;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
